Add slash prefix for local static hrefs
Further to commit fec6da6e, it turns out that the href attributes
for local resources do need to begin with a slash.  For a 404,
the URL is not rewritten, so if the path is, say, abc/def/ghi, an
href beginning "style/..." would be relative to "abc/def", which
is not what we want ....

Put a slash at the beginning of all those href values.